File upload download

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const { Product } = require("../models/product");
 const { validationResult } = require("express-validator");
+const deleteFile = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   //route protection
@@ -18,11 +19,11 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const errors = validationResult(req);
-  if (!errors.isEmpty()) {
+  if (!errors.isEmpty() || !image) {
     return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/edit-product",
@@ -30,14 +31,17 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title,
-        imageUrl,
         price,
         description,
       },
-      errorMessage: errors.array()[0].msg,
-      validationErrors: errors.array(),
+      errorMessage: !image
+        ? "Attatched file is not an image."
+        : errors.array()[0].msg,
+      validationErrors: !image ? [] : errors.array(),
     });
   }
+
+  const imageUrl = image.path;
   const product = new Product({
     title,
     price,
@@ -94,7 +98,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req);
@@ -107,7 +111,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId,
@@ -125,13 +128,17 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save().then((result) => {
         console.log("UPDATED PRODUCT!");
         res.redirect("/admin/products");
       });
     })
     .catch((err) => {
+      console.log("UPDATE Failed!", { err });
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
@@ -158,7 +165,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const pId = req.body.productId;
-  Product.deleteOne({ _id: pId, userId: req.user._id })
+  Product.findById(pId)
+    .then((prod) => {
+      if (!prod) {
+        return next(new Error("Product not Found"));
+      }
+      deleteFile(prod.imageUrl);
+      return Product.deleteOne({ _id: pId, userId: req.user._id });
+    })
     .then(() => {
       res.redirect("/products");
       console.log("PRODUCT DELETED");
